feat(xo-6): add console fullscreen action
