+++ @xen-orchestra/web/src/pages/host/[id].vue
@@ -2,7 +2,7 @@
   <VtsLoadingHero v-if="!isReady" type="page" />
   <VtsObjectNotFoundHero v-else-if="!host" :id="route.params.id" type="page" />
   <RouterView v-else v-slot="{ Component }">
-    <HostHeader :host />
+    <HostHeader v-if="uiStore.hasUi" :host />
     <component :is="Component" :host />
   </RouterView>
 </template>
@@ -13,12 +13,14 @@ import { useHostStore } from '@/stores/xo-rest-api/host.store'
 import type { XoHost } from '@/types/xo/host.type'
 import VtsLoadingHero from '@core/components/state-hero/VtsLoadingHero.vue'
 import VtsObjectNotFoundHero from '@core/components/state-hero/VtsObjectNotFoundHero.vue'
+import { useUiStore } from '@core/stores/ui.store'
 import { computed } from 'vue'
 import { useRoute } from 'vue-router/auto'
 
 const route = useRoute<'/host/[id]'>()
 
 const { isReady, get } = useHostStore().subscribe()
+const uiStore = useUiStore()
 
 const host = computed(() => get(route.params.id as XoHost['id']))
 </script>
