+++ @xen-orchestra/web/src/pages/vm/[id].vue
@@ -2,7 +2,7 @@
   <VtsLoadingHero v-if="!isReady" type="page" />
   <VtsObjectNotFoundHero v-else-if="!vm" :id="route.params.id" type="page" />
   <RouterView v-else v-slot="{ Component }">
-    <VmHeader :vm />
+    <VmHeader v-if="uiStore.hasUi" :vm />
     <component :is="Component" :vm />
   </RouterView>
 </template>
@@ -13,12 +13,14 @@ import { useVmStore } from '@/stores/xo-rest-api/vm.store'
 import type { XoVm } from '@/types/xo/vm.type'
 import VtsLoadingHero from '@core/components/state-hero/VtsLoadingHero.vue'
 import VtsObjectNotFoundHero from '@core/components/state-hero/VtsObjectNotFoundHero.vue'
+import { useUiStore } from '@core/stores/ui.store'
 import { computed } from 'vue'
 import { useRoute } from 'vue-router/auto'
 
 const route = useRoute<'/vm/[id]'>()
 
 const { isReady, get: getVm } = useVmStore().subscribe()
+const uiStore = useUiStore()
 
 const vm = computed(() => getVm(route.params.id as XoVm['id']))
 </script>
